FIX for place_final
Place final had an incorrect function prototype with an unsigned short for tet instead of unsigned int, essentially deleting the offset info.

+++ fillit.h
@@ -31,7 +31,7 @@ unsigned short	smallest_map(short tetcount);
 int				check_tet(unsigned int *tet);
 void			shift_corner(unsigned int *tet);
 void			ft_error(void);
-void			place_final(unsigned short tet, unsigned short *map);
+void			place_final(unsigned int tet, unsigned short *map);
 int				check_hash(short hash, unsigned int tet,\
 unsigned short *map, unsigned short di);
 int				remove_tet(unsigned int *tet, unsigned short *map,\
